Fixed corner case of corner case: don't see -Á or -Í (hyphen+single uppercase letter) as word tokens

+++ release.sh
@@ -1,3 +1,7 @@
+if [ "$1" = "" ]; then
+   echo "Version name argument missing"
+   exit 1
+fi
 echo "Prepare and upload a new Tokenizer version"
 rm -rf build/*
 python setup.py bdist_wheel --universal
